Hide the D3D11 backend if Windows version is older than Win8
