+++ Common/Common.vcxproj.filters
@@ -73,6 +73,7 @@
     <ClInclude Include="Vulkan\VulkanMemory.h">
       <Filter>Vulkan</Filter>
     </ClInclude>
+    <ClInclude Include="OSVersion.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="stdafx.cpp" />
@@ -135,6 +136,7 @@
     <ClCompile Include="MemArenaWin32.cpp" />
     <ClCompile Include="MemArenaAndroid.cpp" />
     <ClCompile Include="MemArenaDarwin.cpp" />
+    <ClCompile Include="OSVersion.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Crypto">
