+++ UI/GameSettingsScreen.cpp
@@ -44,6 +44,7 @@
 
 #include "Common/KeyMap.h"
 #include "Common/FileUtil.h"
+#include "Common/OSVersion.h"
 #include "Core/Config.h"
 #include "Core/Host.h"
 #include "Core/System.h"
@@ -159,6 +160,13 @@ void GameSettingsScreen::CreateViews() {
 #if !defined(_WIN32)
 	renderingBackendChoice->HideChoice(1);  // D3D9
 	renderingBackendChoice->HideChoice(2);  // D3D11
+#else
+	if (!DoesVersionMatchWindows(6, 2)) {
+		// Hide the D3D11 choice if Windows version is older than Windows 8.
+		// We will later be able to support Windows 7 and Vista (unofficially) as well,
+		// but we currently depend on a texture format that's only available in Win8+.
+		renderingBackendChoice->HideChoice(2);  // D3D11
+	}
 #endif
 #if !defined(_WIN32)
 	// TODO: Add dynamic runtime check for Vulkan support on Android
